implemented most of a dashboard.
